# kontrak/openapi.yaml
openapi: 3.1.0
info:
  title: Satu Muatan API
  description: Sistem konsolidasi muatan & bukti mutu rantai pasok hortikultura.
  version: 0.1.0
paths:
...
  /api/slot:
    get:
      tags:
      - slot
      summary: Daftar Slot
      description: 'Ter-scope per peran (K6): KOPERASI → miliknya; PETANI → slot koperasinya;

        PENERIMA → slot yang tujuannya memuat dirinya.'
      operationId: daftar_slot_api_slot_get
      security:
      - HTTPBearer: []
      parameters:
      - name: status
        in: query
        required: false
        schema:
          anyOf:
          - $ref: '#/components/schemas/StatusSlot'
          - type: 'null'
          title: Status
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/SlotItemOut'
                title: Response Daftar Slot Api Slot Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
    post:
      tags:
      - slot
      summary: Buka Slot
      description: Buka slot baru (§9.3). Server menghitung urutan drop nearest-neighbor
        + jarak_km.
      operationId: buka_slot_api_slot_post
      security:
      - HTTPBearer: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/SlotCreate'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlotDetailOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/slot/pratinjau:
    post:
      tags:
      - slot
      summary: Pratinjau Slot
      description: 'Pratinjau §9.3: jarak rute + tabel harga/kg pada berbagai skenario
        volume.'
      operationId: pratinjau_slot_api_slot_pratinjau_post
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PratinjauSlotRequest'
        required: true
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/PratinjauSlotResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
      security:
      - HTTPBearer: []
  /api/slot/{slot_id}:
    get:
      tags:
      - slot
      summary: Detail Slot
      description: 'LAYAR UTAMA DEMO (§9.4). Dipoll 3 detik — harga berjalan, partisipasi,

        rencana armada, waktu_server.'
      operationId: detail_slot_api_slot__slot_id__get
      security:
      - HTTPBearer: []
      parameters:
      - name: slot_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Slot Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlotDetailOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/slot/{slot_id}/gabung:
    post:
      tags:
      - slot
      summary: Gabung Slot
      description: '''Ikut kirim'' — mengunci harga_atap_per_kg petani (tidak pernah
        berubah).'
      operationId: gabung_slot_api_slot__slot_id__gabung_post
      security:
      - HTTPBearer: []
      parameters:
      - name: slot_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Slot Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GabungRequest'
      responses:
        '201':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GabungResponse'
        '409':
          description: LUAPAN_KAPASITAS (§5.5) — dialog dua pilihan
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LuapanKapasitasOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/slot/{slot_id}/gabung/pratinjau:
    post:
      tags:
      - slot
      summary: Pratinjau Gabung
      description: 'Peringatan dini sebelum submit: atap, harga berjalan baru, potensi
        luapan.'
      operationId: pratinjau_gabung_api_slot__slot_id__gabung_pratinjau_post
      security:
      - HTTPBearer: []
      parameters:
      - name: slot_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Slot Id
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/GabungPratinjauRequest'
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/GabungPratinjauResponse'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/slot/{slot_id}/tutup:
    post:
      tags:
      - slot
      summary: Tutup Slot
      description: 'Cutoff (§5.4): tetapkan harga final + jaminan atap, kunci rencana
        armada,

        buat lot per partisipasi (alokasi penerima — K6), pesan ke vendor.'
      operationId: tutup_slot_api_slot__slot_id__tutup_post
      security:
      - HTTPBearer: []
      parameters:
      - name: slot_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Slot Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlotDetailOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/slot/{slot_id}/batal:
    post:
      tags:
      - slot
      summary: Batal Slot
      operationId: batal_slot_api_slot__slot_id__batal_post
      security:
      - HTTPBearer: []
      parameters:
      - name: slot_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Slot Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/SlotDetailOut'
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
  /api/slot/{slot_id}/lot:
    get:
      tags:
      - lot
      summary: Daftar Lot Slot
      description: Daftar lot sebuah slot — layar Muat (§9.5).
      operationId: daftar_lot_slot_api_slot__slot_id__lot_get
      security:
      - HTTPBearer: []
      parameters:
      - name: slot_id
        in: path
        required: true
        schema:
          type: string
          format: uuid
          title: Slot Id
      responses:
        '200':
          description: Successful Response
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/LotOut'
                title: Response Daftar Lot Slot Api Slot  Slot Id  Lot Get
        '422':
          description: Validation Error
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HTTPValidationError'
...
components:
  schemas:
...
    GabungPratinjauRequest:
      properties:
        volume_kg:
          type: integer
          exclusiveMinimum: 0.0
          title: Volume Kg
      type: object
      required:
      - volume_kg
      title: GabungPratinjauRequest
    GabungPratinjauResponse:
      properties:
        harga_atap_per_kg:
          type: integer
          title: Harga Atap Per Kg
        harga_berjalan_baru_per_kg:
          type: integer
          title: Harga Berjalan Baru Per Kg
        luapan:
          type: boolean
          title: Luapan
        pesan:
          anyOf:
          - type: string
          - type: 'null'
          title: Pesan
      type: object
      required:
      - harga_atap_per_kg
      - harga_berjalan_baru_per_kg
      - luapan
      title: GabungPratinjauResponse
    GabungRequest:
      properties:
        komoditas_id:
          type: string
          format: uuid
          title: Komoditas Id
        volume_kg:
          type: integer
          exclusiveMinimum: 0.0
          title: Volume Kg
      type: object
      required:
      - komoditas_id
      - volume_kg
      title: GabungRequest
    GabungResponse:
      properties:
        partisipasi:
          $ref: '#/components/schemas/PartisipasiOut'
        harga_atap_per_kg:
          type: integer
          title: Harga Atap Per Kg
      type: object
      required:
      - partisipasi
      - harga_atap_per_kg
      title: GabungResponse
    HTTPValidationError:
      properties:
        detail:
          items:
            $ref: '#/components/schemas/ValidationError'
          type: array
          title: Detail
      type: object
      title: HTTPValidationError
...
    KoperasiOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        nama:
          type: string
          title: Nama
        kode:
          anyOf:
          - type: string
          - type: 'null'
          title: Kode
        desa:
          anyOf:
          - type: string
          - type: 'null'
          title: Desa
        kecamatan:
          anyOf:
          - type: string
          - type: 'null'
          title: Kecamatan
        kabupaten:
          anyOf:
          - type: string
          - type: 'null'
          title: Kabupaten
        alamat_gudang:
          type: string
          title: Alamat Gudang
        lat:
          type: number
          title: Lat
        lng:
          type: number
          title: Lng
      type: object
      required:
      - id
      - nama
      - alamat_gudang
      - lat
      - lng
      title: KoperasiOut
...
    LotOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        kode_qr:
          type: string
          title: Kode Qr
        partisipasi_id:
          type: string
          format: uuid
          title: Partisipasi Id
        nama_petani:
          type: string
          title: Nama Petani
        nama_komoditas:
          type: string
          title: Nama Komoditas
        volume_kg:
          type: integer
          title: Volume Kg
        penerima_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Penerima Id
        nama_penerima:
          anyOf:
          - type: string
          - type: 'null'
          title: Nama Penerima
        berat_aktual_kg:
          anyOf:
          - type: integer
          - type: 'null'
          title: Berat Aktual Kg
        foto_muat:
          anyOf:
          - type: string
          - type: 'null'
          title: Foto Muat
        waktu_muat:
          anyOf:
          - type: string
            format: date-time
          - type: 'null'
          title: Waktu Muat
        catatan_muat:
          anyOf:
          - type: string
          - type: 'null'
          title: Catatan Muat
        cacat_terlihat:
          type: boolean
          title: Cacat Terlihat
      type: object
      required:
      - id
      - kode_qr
      - partisipasi_id
      - nama_petani
      - nama_komoditas
      - volume_kg
      - cacat_terlihat
      title: LotOut
    LuapanKapasitasOut:
      properties:
        kode:
          type: string
          title: Kode
          default: LUAPAN_KAPASITAS
        harga_baru_per_kg:
          type: integer
          title: Harga Baru Per Kg
        jumlah_atap_terdampak:
          type: integer
          title: Jumlah Atap Terdampak
        slot_alternatif_id:
          anyOf:
          - type: string
            format: uuid
          - type: 'null'
          title: Slot Alternatif Id
        pesan:
          type: string
          title: Pesan
      type: object
      required:
      - harga_baru_per_kg
      - jumlah_atap_terdampak
      - pesan
      title: LuapanKapasitasOut
      description: Body 409 saat gabung memicu luapan (§5.5, K6) — bahan dialog dua
        pilihan.
...
    PartisipasiOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        petani_id:
          type: string
          format: uuid
          title: Petani Id
        nama_petani:
          type: string
          title: Nama Petani
        komoditas_id:
          type: string
          format: uuid
          title: Komoditas Id
        nama_komoditas:
          type: string
          title: Nama Komoditas
        volume_kg:
          type: integer
          title: Volume Kg
        harga_atap_per_kg:
          type: integer
          title: Harga Atap Per Kg
        harga_final_per_kg:
          anyOf:
          - type: integer
          - type: 'null'
          title: Harga Final Per Kg
        kembalian_rp:
          type: integer
          title: Kembalian Rp
        status:
          $ref: '#/components/schemas/StatusPartisipasi'
        bergabung_pada:
          type: string
          format: date-time
          title: Bergabung Pada
      type: object
      required:
      - id
      - petani_id
      - nama_petani
      - komoditas_id
      - nama_komoditas
      - volume_kg
      - harga_atap_per_kg
      - kembalian_rp
      - status
      - bergabung_pada
      title: PartisipasiOut
...
    PratinjauSlotRequest:
      properties:
        tujuan:
          items:
            type: string
            format: uuid
          type: array
          minItems: 1
          title: Tujuan
        skenario_volume:
          items:
            type: integer
          type: array
          minItems: 1
          title: Skenario Volume
          description: mis. [300, 800, 2000]
      type: object
      required:
      - tujuan
      - skenario_volume
      title: PratinjauSlotRequest
    PratinjauSlotResponse:
      properties:
        jarak_km:
          type: number
          title: Jarak Km
        rute:
          items:
            $ref: '#/components/schemas/RuteSegmenOut'
          type: array
          title: Rute
        tabel_harga:
          items:
            $ref: '#/components/schemas/SkenarioHargaOut'
          type: array
          title: Tabel Harga
      type: object
      required:
      - jarak_km
      - rute
      - tabel_harga
      title: PratinjauSlotResponse
    RencanaArmadaOut:
      properties:
        tier:
          items:
            $ref: '#/components/schemas/TierRingkasOut'
          type: array
          title: Tier
        biaya_total:
          type: integer
          title: Biaya Total
        kapasitas_total_kg:
          type: integer
          title: Kapasitas Total Kg
      type: object
      required:
      - tier
      - biaya_total
      - kapasitas_total_kg
      title: RencanaArmadaOut
      description: Rencana armada saat ini — sumber denominator bar kapasitas & petunjuk
        naik kelas tier (§9.4 butir 4).
    RuteSegmenOut:
      properties:
        urutan:
          type: integer
          title: Urutan
        penerima_id:
          type: string
          format: uuid
          title: Penerima Id
        nama_penerima:
          type: string
          title: Nama Penerima
        jarak_segmen_km:
          type: number
          title: Jarak Segmen Km
      type: object
      required:
      - urutan
      - penerima_id
      - nama_penerima
      - jarak_segmen_km
      title: RuteSegmenOut
...
    SkenarioHargaOut:
      properties:
        volume_kg:
          type: integer
          title: Volume Kg
        harga_per_kg:
          type: integer
          title: Harga Per Kg
        biaya_total:
          type: integer
          title: Biaya Total
        kendaraan:
          items:
            type: string
          type: array
          title: Kendaraan
      type: object
      required:
      - volume_kg
      - harga_per_kg
      - biaya_total
      - kendaraan
      title: SkenarioHargaOut
    SlotCreate:
      properties:
        tanggal_kirim:
          type: string
          format: date
          title: Tanggal Kirim
        cutoff_at:
          type: string
          format: date-time
          title: Cutoff At
        tujuan:
          items:
            type: string
            format: uuid
          type: array
          minItems: 1
          title: Tujuan
          description: penerima_id, urutan drop dihitung server (nearest-neighbor)
        permintaan_ids:
          items:
            type: string
            format: uuid
          type: array
          title: Permintaan Ids
          description: permintaan yang hendak dipenuhi slot ini (K6)
      type: object
      required:
      - tanggal_kirim
      - cutoff_at
      - tujuan
      title: SlotCreate
    SlotDetailOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        kode:
          type: string
          title: Kode
        status:
          $ref: '#/components/schemas/StatusSlot'
        tanggal_kirim:
          type: string
          format: date
          title: Tanggal Kirim
        cutoff_at:
          type: string
          format: date-time
          title: Cutoff At
        waktu_server:
          type: string
          format: date-time
          title: Waktu Server
        jarak_km:
          type: number
          title: Jarak Km
        koperasi:
          $ref: '#/components/schemas/KoperasiOut'
        tujuan:
          items:
            $ref: '#/components/schemas/RuteSegmenOut'
          type: array
          title: Tujuan
        volume_total_kg:
          type: integer
          title: Volume Total Kg
        harga_berjalan_per_kg:
          anyOf:
          - type: integer
          - type: 'null'
          title: Harga Berjalan Per Kg
        rencana_saat_ini:
          anyOf:
          - $ref: '#/components/schemas/RencanaArmadaOut'
          - type: 'null'
        partisipasi:
          items:
            $ref: '#/components/schemas/PartisipasiOut'
          type: array
          title: Partisipasi
        atap_saya_per_kg:
          anyOf:
          - type: integer
          - type: 'null'
          title: Atap Saya Per Kg
        hemat_saya_per_kg:
          anyOf:
          - type: integer
          - type: 'null'
          title: Hemat Saya Per Kg
        biaya_total:
          anyOf:
          - type: integer
          - type: 'null'
          title: Biaya Total
        harga_final_per_kg:
          anyOf:
          - type: integer
          - type: 'null'
          title: Harga Final Per Kg
        subsidi_koperasi:
          type: integer
          title: Subsidi Koperasi
          default: 0
      type: object
      required:
      - id
      - kode
      - status
      - tanggal_kirim
      - cutoff_at
      - waktu_server
      - jarak_km
      - koperasi
      - tujuan
      - volume_total_kg
      - partisipasi
      title: SlotDetailOut
      description: Layar utama demo (§9.4). waktu_server = penangkal selisih jam perangkat.
    SlotItemOut:
      properties:
        id:
          type: string
          format: uuid
          title: Id
        kode:
          type: string
          title: Kode
        tanggal_kirim:
          type: string
          format: date
          title: Tanggal Kirim
        cutoff_at:
          type: string
          format: date-time
          title: Cutoff At
        status:
          $ref: '#/components/schemas/StatusSlot'
        jarak_km:
          type: number
          title: Jarak Km
        volume_terkunci_kg:
          type: integer
          title: Volume Terkunci Kg
        kapasitas_rencana_kg:
          anyOf:
          - type: integer
          - type: 'null'
          title: Kapasitas Rencana Kg
        tier_ringkas:
          anyOf:
          - type: string
          - type: 'null'
          title: Tier Ringkas
        jumlah_petani:
          type: integer
          title: Jumlah Petani
      type: object
      required:
      - id
      - kode
      - tanggal_kirim
      - cutoff_at
      - status
      - jarak_km
      - volume_terkunci_kg
      - jumlah_petani
      title: SlotItemOut
      description: Item daftar slot (Beranda §9.2) — bar kapasitas butuh numerator+denominator
        di sini.
    StatusPartisipasi:
      type: string
      enum:
      - TERDAFTAR
      - TERKUNCI
      - DIMUAT
      - SELESAI
      - BATAL
      title: StatusPartisipasi
...
    StatusSlot:
      type: string
      enum:
      - DIBUKA
      - TERKUNCI
      - DIMUAT
      - JALAN
      - SELESAI
      - BATAL
      title: StatusSlot
...
    TierRingkasOut:
      properties:
        kode:
          type: string
          title: Kode
        nama:
          type: string
          title: Nama
        kapasitas_kg:
          type: integer
          title: Kapasitas Kg
      type: object
      required:
      - kode
      - nama
      - kapasitas_kg
      title: TierRingkasOut
...
    ValidationError:
      properties:
        loc:
          items:
            anyOf:
            - type: string
            - type: integer
          type: array
          title: Location
        msg:
          type: string
          title: Message
        type:
          type: string
          title: Error Type
        input:
          title: Input
        ctx:
          type: object
          title: Context
      type: object
      required:
      - loc
      - msg
      - type
      title: ValidationError
  securitySchemes:
    HTTPBearer:
      type: http
      scheme: bearer